changed structure
Now I have stuff in index instead of root, the folder structure is completly different
