+++ app/routes/snippets/$snipId/index.jsx
@@ -1,7 +1,7 @@
 import { useLoaderData, Link, useFormAction, redirect } from "remix";
 import connectDb from "~/db/connectDb.server.js";
 import React, { useState } from 'react';
-import { requireUserSession } from "~/sessions";
+import { requireUserSession } from "~/sessions.server";
 
 
 export async function loader({ params, request }) {
