+++ app/sessions.server.js
@@ -1,4 +1,4 @@
-import { sessionCookie } from "./cookies"
+import { sessionCookie } from "./cookies.server"
 import { createCookieSessionStorage, redirect } from "remix";
 
 const { getSession, commitSession, destroySession } =
